Move types from CustomerLogin to types file
