+++ src/components/App/App.tsx
@@ -20,7 +20,7 @@ import CustomerOrders from '../CustomerOrders/CustomerOrders';
 import CustomerSettings from '../CustomerSettings/CustomerSettings';
 import VendorSettings from '../VendorSettings/VendorSettings';
 import { NewItem, Item } from '../VendorDashboard/VendorDashboard';
-import { Customer } from '../CustomerLogIn/CustomerLogIn';
+import { Customer } from '../../types';
 import { Vendor } from '../VendorLogIn/VendorLogIn';
 import Map from '../Map/Map';
 import ErrorPage from '../ErrorPage/ErrorPage';
@@ -35,8 +35,8 @@ function App() {
   const [isVendor, setIsVendor] = useState<boolean>(false);
   const [currentUserId, setCurrentUserId] = useState<string>('');
   const [appError, setAppError] = useState<string>('');
-  const [selectedZipcode, setSelectedZipcode] = useState<string>('80206')
-  const [selectedRadius, setSelectedRadius] = useState<string>('10')
+  const [selectedZipcode, setSelectedZipcode] = useState<string>('80206');
+  const [selectedRadius, setSelectedRadius] = useState<string>('10');
 
   useEffect(() => {
     const fetchData = async () => {
@@ -62,9 +62,9 @@ function App() {
     setAllVendors([...allVendors, newVendor]);
   }
 
-  function addZipAndRadius(zipcode:string, radius:string) {
-    setSelectedZipcode(zipcode)
-    setSelectedRadius(radius)
+  function addZipAndRadius(zipcode: string, radius: string) {
+    setSelectedZipcode(zipcode);
+    setSelectedRadius(radius);
   }
 
   return appError ? (
@@ -115,7 +115,8 @@ function App() {
           path="/vendordashboard/:vendorid"
           element={
             <VendorDashboard
-              selectedZipcode={selectedZipcode} selectedRadius={selectedRadius}
+              selectedZipcode={selectedZipcode}
+              selectedRadius={selectedRadius}
               allItems={allItems}
               allVendors={allVendors}
               isVendor={isVendor}
@@ -144,8 +145,9 @@ function App() {
             <VendorOrders
               selectedZipcode={selectedZipcode}
               selectedRadius={selectedRadius}
-              isVendor={isVendor} 
-              currentUserId={currentUserId} />
+              isVendor={isVendor}
+              currentUserId={currentUserId}
+            />
           }
         />
         <Route
@@ -154,8 +156,9 @@ function App() {
             <CustomerOrders
               selectedZipcode={selectedZipcode}
               selectedRadius={selectedRadius}
-              isVendor={isVendor} 
-              currentUserId={currentUserId} />
+              isVendor={isVendor}
+              currentUserId={currentUserId}
+            />
           }
         />
         <Route
@@ -175,16 +178,23 @@ function App() {
             <VendorSettings
               selectedZipcode={selectedZipcode}
               selectedRadius={selectedRadius}
-              isVendor={isVendor} 
-              currentUserId={currentUserId} />
+              isVendor={isVendor}
+              currentUserId={currentUserId}
+            />
           }
         />
-        <Route path="/map/:zip/:radius" element={<Map
+        <Route
+          path="/map/:zip/:radius"
+          element={
+            <Map
               addZipAndRadius={addZipAndRadius}
               selectedZipcode={selectedZipcode}
               selectedRadius={selectedRadius}
               isVendor={isVendor}
-              currentUserId={currentUserId}/>} />
+              currentUserId={currentUserId}
+            />
+          }
+        />
         <Route
           path="*"
           element={
